Problem loading more than 32 frames.  Fixed with appropriate casting.
Former-commit-id: d5cedf042e1a2833ce526576110b8e4d404810f0

+++ ios/library/WhirlyGlobeLib/include/Quadtree.h
@@ -299,7 +299,7 @@ protected:
     /// Add an entry for the given flag index
     void addFrameLoaded(int frame);
     /// Clear the flag counts for the given flag entries
-    void clearFlagCounts(int frameFlags);
+    void clearFlagCounts(long long frameFlags);
 
     Mbr mbr;
     int minLevel,maxLevel;
